refactor: move fused QKV dequant from Pro model to WeightLoader
Move _dequant_fused_qkv and _infer_qkv_shards from mimo_v2_pro.py into
WeightLoader methods, keeping all quant logic centralized in weight_utils.
Pro model reduced from 432 to 246 lines.
